Fixes #79 -- Don't show "Delete all" button for events on dashboard

+++ libraries/events.php
@@ -45,11 +45,13 @@ function list_events_table($limit) {
           echo "</tr>";
         }
         echo "</tbody>
-      </table>
-      <div class='pull-right'>
+      </table>";
+      if($limit == 0) {
+      echo "<div class='pull-right'>
         <a class='btn btn-primary' href='delete_event.php?id=all&token=$token'>Delete All</a>
-      </div>
-    </div>";
+      </div>";
+      }
+    echo "</div>";
   } else {
     echo "<center>No events found</center>";
   }
